requirements.txt
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
Django>=4.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0